// core/user-dashboard/store/comment/mutation-types.ts
export enum MutationTypes {
  SET_COMMENT = "SET_COMMENT",
  SET_COMMENTS = "SET_COMMENTS",
  UPDATE_COMMENT_DATA = "UPDATE_COMMENT_DATA",
  UPDATE_NEW_COMMENT_DATA = "UPDATE_NEW_COMMENT_DATA",
  UPDATE_NEW_REPLY_COMMENT_DATA = "UPDATE_NEW_REPLY_COMMENT_DATA",
  UPDATE_EDITING_COMMENT_DATA = "UPDATE_EDITING_COMMENT_DATA",
  SET_LOADING = "SET_LOADING",
  SET_COMMENT_PAGINATION = "SET_COMMENT_PAGINATION",
  UPDATE_COMMENTS_DATA = "UPDATE_COMMENTS_DATA",
}